pyproject.toml
[project]
name = "efs-restore"
version = "0.1.3"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.40.55",
    "hyperscale-ozone>=0.17.2",
    "ozone>=0.1.0",
]
[dependency-groups]
dev = [
  "ruff"
]
tests = [
  "pytest",
  "pytest-cov",
  "pytest-mock"
]
pre-commit = [
  "pre-commit",
  "pre-commit-uv"
]
typing = [
  "mypy",
  "pyright"
]

[tool.uv]
default-groups = ["dev", "pre-commit", "tests", "typing"]


[tool.mypy]
python_version = "3.10"
show_error_codes = true
pretty = true
strict = true

[tool.pyright]
pythonVersion = "3.10"
typeCheckingMode = "basic"

[tool.ruff]
fix = true
show-fixes = true
output-format = "full"

[tool.ruff.lint]
select = [
    "B",  # flake8-bugbear
    "E",  # pycodestyle error
    "F",  # pyflakes
    "I",  # isort
    "UP",  # pyupgrade
    "W",  # pycodestyle warning
]

[tool.ruff.lint.isort]
force-single-line = true
order-by-type = false
